# mumble: Switch to upstream build system: ninja