@dwt 2025-06-06
[Regarding ioq3-scion, I hope to address that upstream here](https://github.com/lschulz/pan-bindings/issues/3)