@dwt 2025-06-19
I found this a great excuse to try out [hyperfine](https://github.com/sharkdp/hyperfine). This is what I got:

```shell
# without patch
❯ nx hyperfine 'nix build .#mumble --rebuild' --warmup 1 --ignore-failure
Benchmark 1: nix build .#mumble --rebuild
  Time (mean ± σ):     73.287 s ±  2.462 s    [User: 0.679 s, System: 3.338 s]
  Range (min … max):   69.478 s … 77.686 s    10 runs
 
  Warning: Ignoring non-zero exit code.
 
# with patch
❯ nx hyperfine 'nix build .#mumble --rebuild' --warmup 1 --ignore-failure
Benchmark 1: nix build .#mumble --rebuild
  Time (mean ± σ):     60.481 s ±  2.828 s    [User: 0.089 s, System: 1.173 s]
  Range (min … max):   57.687 s … 66.910 s    10 runs
 
  Warning: Ignoring non-zero exit code.
```

I see a noticeable improvement, however it's not gigantic.