@dwt 2025-07-12
Friendly ping @felixsinger. Do you think there is enough information in this pull request to warrant it's merging?